requirements.txt
# Core
numpy
pandas
libsumo


# GUI & Plotting
matplotlib
scipy
shapely
networkx

# Reinforcement Learning
gymnasium
stable-baselines3
sb3-contrib
shimmy>=2.0.0
tensorboard
torch
onnx
onnxscript
torchvision

# Config
pyyaml

# Logging
wandb
